Move all public methods to ``_MessageStateHandler``
