+++ pylint/lint/pylinter.py
@@ -25,9 +25,6 @@ from pylint.checkers.base_checker import BaseChecker
 from pylint.config.arguments_manager import _ArgumentsManager
 from pylint.constants import (
     MAIN_CHECKER_NAME,
-    MSG_STATE_CONFIDENCE,
-    MSG_STATE_SCOPE_CONFIG,
-    MSG_STATE_SCOPE_MODULE,
     MSG_TYPES,
     MSG_TYPES_STATUS,
     WarningScope,
@@ -68,9 +65,9 @@ from pylint.utils.pragma_parser import (
 )
 
 if sys.version_info >= (3, 8):
-    from typing import Literal, Protocol
+    from typing import Protocol
 else:
-    from typing_extensions import Literal, Protocol
+    from typing_extensions import Protocol
 
 
 MANAGER = astroid.MANAGER
@@ -306,15 +303,6 @@ class PyLinter(
             ("Messages control", "Options controlling analysis messages"),
             ("Reports", "Options related to output formatting and reporting"),
         )
-        self._options_methods = {
-            "enable": self.enable,
-            "disable": self.disable,
-            "disable-next": self.disable_next,
-        }
-        self._bw_options_methods = {
-            "disable-msg": self._options_methods["disable"],
-            "enable-msg": self._options_methods["enable"],
-        }
         self.fail_on_symbols: list[str] = []
         """List of message symbols on which pylint should fail, set by --fail-on."""
         self._error_mode = False
@@ -500,16 +488,6 @@ class PyLinter(
     def any_fail_on_issues(self) -> bool:
         return any(x in self.fail_on_symbols for x in self.stats.by_msg.keys())
 
-    def disable_noerror_messages(self) -> None:
-        for msgcat, msgids in self.msgs_store._msgs_by_category.items():
-            # enable only messages with 'error' severity and above ('fatal')
-            if msgcat in {"E", "F"}:
-                for msgid in msgids:
-                    self.enable(msgid)
-            else:
-                for msgid in msgids:
-                    self.disable(msgid)
-
     def disable_reporters(self) -> None:
         """Disable all reporters."""
         for _reporters in self._reports.values():
@@ -529,26 +507,6 @@ class PyLinter(
         self.set_option("reports", False)
         self.set_option("persistent", False)
         self.set_option("score", False)
-
-    def list_messages_enabled(self) -> None:
-        emittable, non_emittable = self.msgs_store.find_emittable_messages()
-        enabled = []
-        disabled = []
-        for message in emittable:
-            if self.is_message_enabled(message.msgid):
-                enabled.append(f"  {message.symbol} ({message.msgid})")
-            else:
-                disabled.append(f"  {message.symbol} ({message.msgid})")
-        print("Enabled messages:")
-        for msg in enabled:
-            print(msg)
-        print("\nDisabled messages:")
-        for msg in disabled:
-            print(msg)
-        print("\nNon-emittable messages with current interpreter:")
-        for msg_def in non_emittable:
-            print(f"  {msg_def.symbol} ({msg_def.msgid})")
-        print("")
 
     # block level option handling #############################################
     # see func_block_disable_msg.py test case for expected behaviour
@@ -1162,89 +1120,6 @@ class PyLinter(
             self.reporter.display_reports(sect)
         return note
 
-    # Adding (ignored) messages to the Message Reporter
-
-    def _get_message_state_scope(
-        self,
-        msgid: str,
-        line: int | None = None,
-        confidence: interfaces.Confidence | None = None,
-    ) -> Literal[0, 1, 2] | None:
-        """Returns the scope at which a message was enabled/disabled."""
-        if confidence is None:
-            confidence = interfaces.UNDEFINED
-        if confidence.name not in self.config.confidence:
-            return MSG_STATE_CONFIDENCE  # type: ignore[return-value] # mypy does not infer Literal correctly
-        try:
-            if line in self.file_state._module_msgs_state[msgid]:
-                return MSG_STATE_SCOPE_MODULE  # type: ignore[return-value]
-        except (KeyError, TypeError):
-            return MSG_STATE_SCOPE_CONFIG  # type: ignore[return-value]
-        return None
-
-    def _is_one_message_enabled(self, msgid: str, line: int | None) -> bool:
-        """Checks state of a single message for the current file.
-
-        This function can't be cached as it depends on self.file_state which can
-        change.
-        """
-        if line is None:
-            return self._msgs_state.get(msgid, True)
-        try:
-            return self.file_state._module_msgs_state[msgid][line]
-        except KeyError:
-            # Check if the message's line is after the maximum line existing in ast tree.
-            # This line won't appear in the ast tree and won't be referred in
-            # self.file_state._module_msgs_state
-            # This happens for example with a commented line at the end of a module.
-            max_line_number = self.file_state.get_effective_max_line_number()
-            if max_line_number and line > max_line_number:
-                fallback = True
-                lines = self.file_state._raw_module_msgs_state.get(msgid, {})
-
-                # Doesn't consider scopes, as a 'disable' can be in a
-                # different scope than that of the current line.
-                closest_lines = reversed(
-                    [
-                        (message_line, enable)
-                        for message_line, enable in lines.items()
-                        if message_line <= line
-                    ]
-                )
-                _, fallback_iter = next(closest_lines, (None, None))
-                if fallback_iter is not None:
-                    fallback = fallback_iter
-
-                return self._msgs_state.get(msgid, fallback)
-            return self._msgs_state.get(msgid, True)
-
-    def is_message_enabled(
-        self,
-        msg_descr: str,
-        line: int | None = None,
-        confidence: interfaces.Confidence | None = None,
-    ) -> bool:
-        """Return whether this message is enabled for the current file, line and confidence level.
-
-        This function can't be cached right now as the line is the line of
-        the currently analysed file (self.file_state), if it changes, then the
-        result for the same msg_descr/line might need to change.
-
-        :param msg_descr: Either the msgid or the symbol for a MessageDefinition
-        :param line: The line of the currently analysed file
-        :param confidence: The confidence of the message
-        """
-        if confidence and confidence.name not in self.config.confidence:
-            return False
-        try:
-            msgids = self.msgs_store.message_id_store.get_active_msgids(msg_descr)
-        except exceptions.UnknownMessageError:
-            # The linter checks for messages that are not registered
-            # due to version mismatch, just treat them as message IDs
-            # for now.
-            msgids = [msg_descr]
-        return any(self._is_one_message_enabled(msgid, line) for msgid in msgids)
-
     def _add_one_message(
         self,
         message_definition: MessageDefinition,
@@ -1399,50 +1274,3 @@ class PyLinter(
                 message_definition.msgid,
                 line,
             )
-
-    # Setting the state (disabled/enabled) of messages and registering them
-
-    def disable(
-        self,
-        msgid: str,
-        scope: str = "package",
-        line: int | None = None,
-        ignore_unknown: bool = False,
-    ) -> None:
-        """Disable a message for a scope."""
-        self._set_msg_status(
-            msgid, enable=False, scope=scope, line=line, ignore_unknown=ignore_unknown
-        )
-        self._register_by_id_managed_msg(msgid, line)
-
-    def disable_next(
-        self,
-        msgid: str,
-        scope: str = "package",
-        line: int | None = None,
-        ignore_unknown: bool = False,
-    ) -> None:
-        """Disable a message for the next line."""
-        if not line:
-            raise exceptions.NoLineSuppliedError
-        self._set_msg_status(
-            msgid,
-            enable=False,
-            scope=scope,
-            line=line + 1,
-            ignore_unknown=ignore_unknown,
-        )
-        self._register_by_id_managed_msg(msgid, line + 1)
-
-    def enable(
-        self,
-        msgid: str,
-        scope: str = "package",
-        line: int | None = None,
-        ignore_unknown: bool = False,
-    ) -> None:
-        """Enable a message for a scope."""
-        self._set_msg_status(
-            msgid, enable=True, scope=scope, line=line, ignore_unknown=ignore_unknown
-        )
-        self._register_by_id_managed_msg(msgid, line, is_disabled=False)
